[Refactor] Move old modifier to legacy/stakemodifier.*
