+++ src/kernel.h
@@ -75,9 +75,4 @@ bool CheckProofOfStake(const CBlock& block, std::string& strError, const CBlockI
  */
 bool GetStakeKernelHash(uint256& hashRet, const CBlock& block, const CBlockIndex* pindexPrev = nullptr);
 
-
-/* Old Stake Modifier */
-bool GetOldStakeModifier(CStakeInput* stake, uint64_t& nStakeModifier);
-bool ComputeNextStakeModifier(const CBlockIndex* pindexPrev, uint64_t& nStakeModifier, bool& fGeneratedStakeModifier);
-
 #endif // PIVX_KERNEL_H
